Conditional Rendering

+++ src/main/java/org/pahappa/systems/requisitionapp/views/BudgetLineCategoryManagedBean.java
@@ -4,6 +4,7 @@ import org.pahappa.systems.requisitionapp.models.BudgetLine;
 import org.pahappa.systems.requisitionapp.models.BudgetLineCategory;
 import org.pahappa.systems.requisitionapp.models.Requisition;
 import org.pahappa.systems.requisitionapp.models.utils.BudgetLineStatus;
+import org.pahappa.systems.requisitionapp.models.utils.Permission;
 import org.pahappa.systems.requisitionapp.models.utils.RequisitionStatus;
 import org.pahappa.systems.requisitionapp.services.BudgetLineCategoryService;
 import org.pahappa.systems.requisitionapp.services.BudgetLineService;
@@ -73,6 +74,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
     }
 
     public void addBudgetLineToBudgetLineCategory() {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.CREATE_BUDGET_LINE)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         BudgetLineCategory budgetLineCategory = budgetLineCategoryService.getBudgetLineCategoryByName(budgetLineCategoryName);
         try {
             if (budgetLineCategory != null) {
@@ -92,6 +98,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
     }
 
     public void updateBudgetLine() {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.EDIT_BUDGET_LINE)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         try {
             if (!selectedBudgetLine.getStatus().equals(BudgetLineStatus.DRAFT)) {
                 FacesContext.getCurrentInstance().addMessage(null,
@@ -117,6 +128,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
      */
 
     public void approveBudgetLine() {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.APPROVE_BUDGET_LINE)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         if (selectedBudgetLine.getStatus().equals(BudgetLineStatus.DRAFT)) {
             selectedBudgetLine.setStatus(BudgetLineStatus.APPROVED);
             budgetLineService.updateBudgetLine(selectedBudgetLine);
@@ -129,6 +145,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
     }
 
     public void rejectBudgetLine() {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.REJECT_BUDGET_LINE)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         if (selectedBudgetLine.getStatus().equals(BudgetLineStatus.DRAFT)) {
             selectedBudgetLine.setStatus(BudgetLineStatus.REJECTED);
             budgetLineService.updateBudgetLine(selectedBudgetLine);
@@ -141,6 +162,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
 
 
     public void deleteBudgetLine(BudgetLine budgetLine) {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.DELETE_BUDGET_LINE)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         try {
             List<Requisition> budgetLineRequisitions = budgetLine.getRequisitions();
             for (Requisition requisition : budgetLineRequisitions){
@@ -179,6 +205,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
     }
 
     public void addBudgetLineCategory() {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.CREATE_BUDGET_LINE_CATEGORY)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         try {
             budgetLineCategoryService.createBudgetLineCategory(newBudgetLineCategory);
             FacesContext.getCurrentInstance().addMessage(null,
@@ -192,6 +223,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
     }
 
     public void updateBudgetLineCategory() {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.EDIT_BUDGET_LINE_CATEGORY)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         try {
             budgetLineCategoryService.updateBudgetLineCategory(selectedBudgetLineCategory);
             FacesContext.getCurrentInstance().addMessage(null,
@@ -204,6 +240,11 @@ public class BudgetLineCategoryManagedBean implements Serializable {
     }
 
     public void deleteBudgetLineCategory(BudgetLineCategory category) {
+        if (!LoginBean.getCurrentUser().getRole().getPermissions().contains(Permission.DELETE_BUDGET_LINE_CATEGORY)){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "Current user does not have permission to access this function.", null));
+            return;
+        }
         try {
             budgetLineCategoryService.deleteBudgetLineCategory(category);
             FacesContext.getCurrentInstance().addMessage(null,
